cambios

+++ src/main/java/com/copsis/dto/AmortizacionDTO.java
@@ -1,7 +1,5 @@
 package com.copsis.dto;
 
-import java.math.BigDecimal;
-
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 
@@ -11,12 +9,12 @@ import lombok.Data;
 @JsonInclude(Include.NON_NULL)
 public class AmortizacionDTO {
 	private Integer id;
-	private BigDecimal seguroDanos;
-	private BigDecimal aportacionCapital;
-	private BigDecimal capital;
-	private BigDecimal interes;
-	private BigDecimal iva;
-	private BigDecimal pago;
-	private BigDecimal abonoCapital;
-	private BigDecimal capitalNuevo;
+	private Double seguroDanos;
+	private Double aportacionCapital;
+	private Double capital;
+	private Double interes;
+	private Double iva;
+	private Double pago;
+	private Double abonoCapital;
+	private Double capitalNuevo;
 }
